requirements.txt
Django==3.2.8
gunicorn==20.1.0
pandas
djangorestframework
django-filter
django-q